# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Order API",
    "version": "1.0.0"
  },
  "paths": {
    "/notifications/users/{id}": {
      "post": {
        "summary": "Send a notification to a user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          },
          "500": {
            "description": "Failed to send notification"
          }
        }
      }
    },
    "/notifications/users": {
      "post": {
        "summary": "Send a notification to multiple users",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "notification": {
                    "$ref": "#/components/schemas/NotificationCreate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notifications sent successfully"
          },
          "500": {
            "description": "Failed to send notifications"
          }
        }
      }
    },
    "/notifications/subscribe": {
      "post": {
        "summary": "Subscribe to notifications",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscribed to notifications successfully"
          },
          "500": {
            "description": "Failed to subscribe to notifications"
          }
        }
      }
    },
    "/notifications/unsubscribe": {
      "post": {
        "summary": "Unsubscribe from notifications",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsubscribed from notifications successfully"
          },
          "500": {
            "description": "Failed to unsubscribe from notifications"
          }
        }
      }
    },
    "/notifications/users/{id}/history": {
      "get": {
        "summary": "Get notification history for a user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "summary": "Get notifications for a user",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "summary": "Mark a notification as read",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          },
          "404": {
            "description": "Notification not found"
          }
        }
      }
    },
    "/notifications/{id}/unread": {
      "put": {
        "summary": "Mark a notification as unread",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as unread"
          },
          "404": {
            "description": "Notification not found"
          }
        }
      }
    }
  }
}